Refactor Student Profile to use type in api-output
