+++ src/web/app/pages-help/instructor-help-page/instructor-help-students-section/instructor-help-students-section.component.spec.ts
@@ -3,9 +3,8 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { Component, Input } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { Gender, JoinState, Student } from '../../../../types/api-output';
+import { Gender, JoinState, Student, StudentProfile } from '../../../../types/api-output';
 import { SearchStudentsTable } from '../../../pages-instructor/instructor-search-page/instructor-search-page.component';
-import { StudentProfile } from '../../../pages-instructor/student-profile/student-profile';
 import { InstructorHelpStudentsSectionComponent } from './instructor-help-students-section.component';
 
 @Component({ selector: 'tm-example-box', template: '' })
@@ -21,13 +20,13 @@ class StudentResultTableStubComponent {
 @Component({ selector: 'tm-student-profile', template: '' })
 class StudentProfileStubComponent {
   @Input() studentProfile: StudentProfile = {
+    name: '',
     shortName: '',
     email: '',
     institute: '',
     nationality: '',
     gender: Gender.FEMALE,
     moreInfo: '',
-    pictureKey: '',
   };
 }
 @Component({ selector: 'tm-course-related-info', template: '' })
