+++ src/web/app/pages-instructor/instructor-course-student-details-page/instructor-course-student-details-page.component.spec.ts
@@ -3,8 +3,7 @@ import { Component, Input } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatSnackBarModule } from '@angular/material';
 import { RouterTestingModule } from '@angular/router/testing';
-import { Gender, JoinState, Student } from '../../../types/api-output';
-import { StudentProfile } from '../student-profile/student-profile';
+import { Gender, JoinState, Student, StudentProfile } from '../../../types/api-output';
 import {
   InstructorCourseStudentDetailsPageComponent,
 } from './instructor-course-student-details-page.component';
